requirements.txt
django==2.2.24
djangorestframework==3.12.4
markdown
django-filter
djangorestframework-simplejwt
django-oauth-toolkit